ml/requirements.txt
catboost
Levenshtein
pdfminer.six
numpy
pandas